Introduce new build flag for bootld and add EFI boot targets.
EFI needs BootLd to use -BSymbolic and others to link.
Also moved BoardSetup to use that flag instead of Jamfile magic.
BoardSetup was called and added to XX_CC_FLAGS before
XX_CC_FLAGS was set so moved BoardSetup to after it, which also
made sense for this change.

This allow building a runnable(!) EFI app with
	jam -sHAIKU_BOOT_PLATFORM=efi haiku_loader.efi
Atm it only prints a message and waits for a key.