booting: support identifying disks & partitions by UUID.
This adds a UUID field for partition_data, and updates the
GPT partitioning system to add the UUIDs for both the
partitioning system and each child partition.

The VFS boot code has been updated to detect whether disks
and partitions should be identified by UUID, skipping the
fallback mechanisms used for other platforms.

These changes will allow the UEFI boot loader to correctly
pass the boot device and partition to the kernel.